feat: update navigation to bottom navigation bar
