+++ packages/helpwave_localization/lib/l10n/app_en.arb
@@ -124,5 +124,7 @@
   "discharge": "Discharge",
   "addTask": "Add Task",
   "all": "All",
-  "searchPatient": "Search Patient"
+  "searchPatient": "Search Patient",
+  "task": "Task",
+  "create": "Create"
 }
